CAMEL-1588: Configuring a custom default thread pool profile. Added better debug logging for thread pool activity.
git-svn-id: https://svn.apache.org/repos/asf/camel/trunk@923718 13f79535-47bb-0310-9956-ffa450edef68

+++ components/camel-spring/src/main/java/org/apache/camel/spring/CamelThreadPoolFactoryBean.java
@@ -52,7 +52,7 @@ public class CamelThreadPoolFactoryBean extends IdentifiedType implements Factor
     @XmlAttribute
     private Integer maxPoolSize;
     @XmlAttribute
-    private Integer keepAliveTime = 60;
+    private Long keepAliveTime = 60L;
     @XmlAttribute
     @XmlJavaTypeAdapter(TimeUnitAdapter.class)
     private TimeUnit units = TimeUnit.SECONDS;
@@ -120,11 +120,11 @@ public class CamelThreadPoolFactoryBean extends IdentifiedType implements Factor
         this.maxPoolSize = maxPoolSize;
     }
 
-    public Integer getKeepAliveTime() {
+    public Long getKeepAliveTime() {
         return keepAliveTime;
     }
 
-    public void setKeepAliveTime(Integer keepAliveTime) {
+    public void setKeepAliveTime(Long keepAliveTime) {
         this.keepAliveTime = keepAliveTime;
     }
